Tool Change Macro
This macro adds a 'Tool Length' button to the Control menu. It This allows you to use one tool as a reference and then to set the Z offset of the WCS to maintain the same tool height through tool changes.

It uses the GCODE preset locations G30 for the probe position and G28 for a preset tool change position (hence the machine needs to be homed) and these locations need to be set before using the macro (**IMPORTANT!**). Using these locations has the advantage that if you want to adjust (e.g.) your probe position, you can just jog the machine to where you want it to start and send 'G30.1' (or G28.1 for the tool change location) and it's done without needing to edit any code. If you already use G28 for something, then just ignore the 'Go To Tool Change' button.

After use, it gives the option of returning the tool to the location it was in when the dialog was opened - use with caution!

Any changes in Z zero between tool changes are carried through to the new tool.

There are some customisable settings near the start of the file to save you from going digging for probe speeds, etc.

The macro to checks the 'homedRecently' status in Control and warns if the machine needs homing. Unfortunately, at the time of writing, the homedRecently status shows that the machine needs homing after every error (e.g. syntax error in gcode file, attempting to jog beyond the machine limits, etc.) even though the position is reliable. Because of this, I've given the option to carry on regardless, and also the option to disable the check in the macro (set tc_ignorehomestatus to "yes").

Known issue: If the probe isn't in the expected state before probing, or the requested probe travel is > machine travel these generate hard errors that my macro can't trap. If they happen, you will need to close the dialog and re-open it. Similarly if the requested probe travel exceeds the machine travel - take care not to set G30 too low.

It isn't possible to pause a program to use this macro, as (I think) the comms port isn't released while the program is paused, so each tool will need its separate file.