Laser engraving test pattern generator
Running this macro generates the GCode to produce a matrix of test patches with a range of power and feed rate settings to determine the optimum parameters for LASER marking.

Each row has a different power level, and each column has a different feed rate.The parameters can be changed by editing the first few lines of the macro:

Resolution (number of lines per mm);
Maximum and minimum feed rates and power levels;
Number of rows / columns;
The height and width of each patch, and the size of the gap between them.

Power settings use splindle speed units (grbl default is 0-1000). The value in $30 for the maximum speed is what you need to put into the macro to get 100% power.

The labels are engraved at the half way point of both power and speed settings.

The label text is fixed at 3mm high which will mean that the labels overwrite each other at patch sizes less than about 3mm.

When the macro is run, any loaded GCode will be replaced by the output of the macro. The GCode can be run or saved as required.

The code uses 'M4' to control the LASER which requires Laser Mode to be enabled in grbl ($32=1).