Laser Engraver Focus / Depth of Field Test Pattern
This macro generates the GCode for a test pattern that varies in Z height around the current Z0 to act as an aid in checking / setting LASER focus.

Parameters can be changed by altering the variables at the start of the macro:

The range of movement from the current Z0 - note that this is +/- , so a value of 3(say) gives a range of Z-3 to Z+3;
The power and speed for the pattern (power is a percentage of the spindle units);
The number and spacing of the line pattern - playing with the spacing can really highlight the linewidth broadening when the beam is out of focus.

The GCode generates travel that extends beyond the edge of the pattern to allow the machine to accelerate to the programmed speed. By default, this means that the head will move into negative X coordinates (to the left of X0) to start the pattern at X0. If this causes a problem, set 'xplus_Only' to "yes" - the pattern will be shifted to the right to allow enough travel for the machine to accelerate.

Set 'x_acceleration' to the actual machine X acceleration ($120) for the best results, or to zero to disable the over-travel.

As ever, use at your own risk.